Initial Work Towards 64-bit patch Still likely will have issues if you try to build 64 Filesystem and model code are likely big culprits too

+++ codemp/rd-vanilla/tr_init.cpp
@@ -1295,7 +1295,7 @@ void R_Init( void ) {
 //	Swap_Init();
 
 #ifndef FINAL_BUILD
-	if ( (int)tess.xyz & 15 ) {
+	if ( (intptr_t)tess.xyz & 15 ) {
 		Com_Printf( "WARNING: tess.xyz not 16 byte aligned (%x)\n",(int)tess.xyz & 15 );
 	}
 #endif
